Update patch set 1
Patch Set 1: Code-Review-1

Please keep the code in the same order when moving it from vhost.c to the new vhost_scsi.c file.  Otherwise it makes the diff impossible to review to ensure nothing was changed by accident.  For example, in vhost.c, new_device is defined immediately before destroy_device.  But then in the new vhost_scsi.c file, new_device is both moved to the new file and moved to a different location relative to other functions that were moved.

If function definitions need to be reordered, please do that in a separate patch from the one that moves them between files.

Patch-set: 1
Label: Code-Review=-1