Update patch set 1
Patch Set 1: Code-Review-1

I'm not sure I like the name "stub", or that the application is in app when it is really only used for tests. Maybe naming it "primary", or "spdkd", or "master" would be better? Also, what about putting it in the test directory instead? Do you think this code will get used outside of our test environment?

Patch-set: 1
Label: Code-Review=-1