Update patch set 2
Patch Set 2:

The Auto Test failure looks like due to the DPDK memzone creation failure where the same name existed (from the first test case calsoft.sh, the name is spdk_ring_1 from spdk_ring_create()). It seems that the iscsi_tgt is not properly cleaned up in the calsoft.sh test set. When for the filesystem.sh test set, it tries to create another ring from spdk_ring_create() with same name spdk_ring_1, it failed as this name of memzone already existed (we are now running under the multi-process situation). As long as we can properly cleanup the iscsitgt in the testset to free the memzone, it will be no problem for the next testset.

Patch-set: 2
Reviewer: GangCao <1011207@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Verified=0