Update patch set 1
Patch Set 1:

One more thought - Jim submitted some code to examples/nvme/perf to track latency as a histogram. It's not a general purpose library like yours, but his histogram implementation is extremely clever. Please take a look at it (examples/nvme/perf/perf.c on master) - it eliminates the need for different scales or for the user to set bounds.

Patch-set: 1