Update patch set 3
Patch Set 3: Code-Review-1

I'd like to keep this.  We haven't used it to date because we haven't had a good way to test it.  The plan was to keep LUN reset as a soft reset.  If LUN reset fails (i.e. the client sees that it didn't complete all I/O) then it will escalate to a device reset which will be a bdev HARD reset.

I'd say do your best to keep this functional based on what's there currently.  I'll work with Cunyin to get this implemented and tested as part of the bdev error injection work.

Patch-set: 3
Reviewer: Jim Harris <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1