Update patch set 1
Patch Set 1:

Hey John - thanks for this patch.  This obviously touches a lot of code so will be susceptible to needing rebases.  We will work on getting this patch in next week - we may do some of the rebases here to accelerate that process - your name will still be the author on the patch though.  :-)

Patch-set: 1