doc: update root README.md
First in a series of doc updates. Add in-page links for sections
and various updates to each section for clarity. Note that this
readme is focused on getting the code and running unit tests
but includes some other basic info also. The last section links
to spdk.io where the developer page is the home for details on
contributing.

Change-Id: Ibbe5a070488cf295c1decd1a8b72b23fd140e942
Signed-off-by: Paul Luse <paul.e.luse@intel.com>