doc: update root README.md
First in a series of doc updates. This removes some duplicate info
that we also have on spdk.io and replaces some of the links found
on that site with a link to spdk.io. Also fixes a few minor missing
steps for all OSs and all were tested from a fresh install
by copying and pasting only info from the readme.

Upcoming patches include removing duplicate info from the developer
section of spdk.io and expanding this readme to include simple
walkthroughs for patch creation & submission. The intent is that
the quick and easy steps are documented here and the supporting
information and more complex operations be covered on spdk.io in
the developer tab.

Slight changes made to /configure to be more FreeBSD friendly and
match the readme.

Change-Id: Ibbe5a070488cf295c1decd1a8b72b23fd140e942
Signed-off-by: Paul Luse <paul.e.luse@intel.com>