Update patch set 1
Patch Set 1:

> So far this new func is called no where? Do we have any chance that the accumulated read len and write len will overflow as they are accumulated on each IO?
This will be used by VTune and other monitoring services. Its purpose is to provide stats of time intervals managed by the caller. The stats are reset on each call for the next interval. Yes, if this function is not called (for a very long time,) the counters can overflow, but they are only meaningful after they have been reset.

Patch-set: 1
Label: Verified=0