Update patch set 1
Patch Set 1: Code-Review-1

(3 comments)

Is this library available somewhere else publicly?

I don't think we really want to carry a copy of it in SPDK, since it's not part of SPDK itself - it should just be added as a dependency required to build SPDK with VTune tracing support.

Patch-set: 1
Label: Code-Review=-1