env/pci: add write combined support for PCI BAR memory space
To enabled Controller Memory Buffer feature, the PCI BAR memory
should be remapped as write combined memory space to increase
the performance. It's make more sense to put those APIs into
DPDK library, but before DPDK has support on it, we can keep
it in SPDK library. To test this feature, enabled it by default,
and users can disable it in their own driver's callback.

Change-Id: If4952513d77cecaa4f9403fbea811d86916ee87c
Signed-off-by: Changpeng Liu <changpeng.liu@intel.com>