event: Removed ability to dump config from running instance
While we move from configuration files to RPC as the primary
configuration mode, remove this feature. It had been
badly neglected anyway and wasn't fully functional.

As RPC becomes the primary mechanism for configuration,
we'll need a new tool that can use RPC to query the status
of an application and generate a configuration file.

Change-Id: I62f3ef304701d640629401b59a1f8fae8a51e2f1
Signed-off-by: Ben Walker <benjamin.walker@intel.com>