nvme: allocate queue memory from PCI device socket
Use the socket ID of the NVMe PCIe device to determine where to allocate
memory for the submission and completion queues, as well as trackers,
which contain the SGL/PRP list.

This is the most reasonable default to try to get the best possible
performance on multi-socket systems, although in practice the
differences are very minimal.  DPDK will automatically revert
to use memory on other sockets if there is no memory available on the
specified socket.

Change-Id: I1da1ed4d5fd83030c270aac43ce071847c6eb6c5
Signed-off-by: Daniel Verkamp <daniel.verkamp@intel.com>