nvme: allocate queue memory from PCI device socket
Use the socket ID of the NVMe PCIe device to determine where to allocate
memory for the submission and completion queues, as well as trackers,
which contain the SGL/PRP list.  These are all transferred via DMA by
the NVMe device, and putting them on the same NUMA node as the device
should improve performance on multi-socket systems.

Change-Id: I1da1ed4d5fd83030c270aac43ce071847c6eb6c5
Signed-off-by: Daniel Verkamp <daniel.verkamp@intel.com>