Update patch set 3
Patch Set 3: Code-Review-1

Do we need this change?  This modifies the copy of DPDK rte_vhost in our repository - I was hoping to keep those changes to a minimum if possible (which is why we ignore this directory for astyle for example).

Patch-set: 3
Reviewer: Jim Harris <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1