Update patch set 8
Patch Set 8: Code-Review-1

What does it mean to inject a "reset"? I'm confused about what we're really trying to do here. For read/write/flush/trim, the injection means fail the next I/O of that type. But that's not what we want to do with resets. 

One optional would be to inject a timeout condition. For example, make the vbdev simply not complete the next read/write command (until the user issues a reset).

Another option is to provide a mechanism to directly send a reset to the bdev layer out of band (using JSON RPC) - but that isn't an error injection. That's just calling spdk_bdev_reset in response to an RPC.

Maybe we should do both.

Patch-set: 8
Reviewer: Ben Walker <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1