Update patch set 8
Patch Set 8:

Hi Ben, yes, you are right, for the read/write/flush/trim, what we are trying to do here is just fail them and return error, but for reset, just as your mentioned, we try to hold on the next command in list and wait for the client to send an reset command, this patch is just do the works as your option 1. for the option 2 you mentioned, maybe we could submit one new patch?

Patch-set: 8
Reviewer: Cunyin Chang <1011231@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0