Update patch set 6
Patch Set 6:

I'm generally not opposed to this, but the API in log.h is already just passing a string (the log library could just ignore calls to set the facility if it has no equivalent).

Also, Fujitsu did a bunch of cleanup in this area, so I suspect they are actually using this.

If it's not too much trouble to keep it, I think we should (pending responses from anyone using it).

Patch-set: 6
Label: Verified=0