vhost_scsi: implemented vhost device hotremove
Vhost devices can now be removed via RPC at any time.
When hotremove is triggered, a scsi_dev pointer is first unset in
scsi_dev_w, so that any upcoming IO is instantly rejected. Then,
an eventq event is queued. It will be sent soon after on mgmt_poller
tick or when first IO in requestq is rejected. This ensures that guest
VM is notified quick enough not to issue any scsi reset on the device.

Change-Id: Ib91dd6eb11ae8c777e8b9498d140e89a3fe382db
Signed-off-by: Dariusz Stojaczyk <dariuszx.stojaczyk@intel.com>