Update patch set 4
Patch Set 4:

Using remove_cb() to return the discovery_ctrlr looks like a wrong method. Added a new flag in the ctrlr_opts to handle this more properly.

The background is that the nvme-cli needs to discovery_ctrlr to do the general "get-log" operations to find the ctrlr entries and for the related operations as an entry point.

Patch-set: 4