Update patch set 4
Patch Set 4:

I've tested this change and along with the nvme-cli change (http://spdk.intel.com/gerrit/#/c/8017/). On my testing environment, the nvmf_tgt started with configured dpdk mem size and nvme-cli can also start with configured mem size to discover the subsystems. Here multi-process is not leveraged. And the purpose is to integrate with SPDK NVMe-oF host capability.

Patch-set: 4