Update patch set 2
Patch Set 2: Code-Review-1

I'm not sure we need RPCs for GPT.  It is OK to have this one, but it is really only for testing purposes I think.

To resolve this issue with asynchronous I/O from vbdevs like GPT, how about we add a script function like waitforbdev?  You pass a string, the function calls get_bdevs RPC, and when it sees the bdev, it returns from the function.  It can call it every millisecond and fail out with error if it takes more than 1 seconds (or some relatively large value).

This way we do not need to add extra "sleep 5" in the tests.  These extra delays in the code make our tests run longer and longer over time.  We really want to find programmatic ways to wait until the test can proceed, instead of just adding delays.

So could you do this sequence of patches:

1) my open/close patch
2) new script function waitforbdev
3) your GPT bdev patch (using waitforbdev instead of sleep) - also just use kill -0 $nbd_pid instead of the ps stuff
4) this RPC patch

Patch-set: 2
Label: Code-Review=-1