script: add waitforbdev
The purpose is to control the time for test instead
of using sleep 1 or sheep 5when waiting for the ready
of bdevs.

BTW, we cannot use return value to differentiate,
the reaon is that our script use "set -e". If non-zero
value is detected, it will trap. So use an exported variable
SPDK_BDEV_READY.

Change-Id: I0319bcc270257612077da42024217d532cbc34e3
Signed-off-by: Ziye Yang <optimistyzy@gmail.com>