Update patch set 4
Patch Set 4: -Code-Review

Look like we can have a global value for the tsc_end and in the task_complete() callback, as we already have the spdk_get_ticks() call there, save a local variable to spdk_get_ticks() and compare with the global tsc_end to determine whether completing the IO by another global flag. In the while(1), condition, check that global flag instead of calling spdk_get_ticks() to break out. In this way, no more call of spdk_get_ticks() is needed.

Patch-set: 4