Update patch set 2
Patch Set 2:

The = {0} is fine but I did need to do something different for non-zero struct initializers.  The MOCK_STRUCT_INIT macro is needed because you can't have a { a, b, c } as a parm in a macro as it takes the commas as separators so using VA_ARGS is a trick to get that through. Disregard prev comment on needing to address something in the _V macros, the issue doesn't exist until the next patch anyway...

Patch-set: 2
Label: Code-Review=0