Update patch set 9
Patch Set 9: Code-Review-1

Also there is an issue:

Vbdev_gpt_detection is asynchronous, and current this patch does not address this issue.  When all subsystem is ready, bdevperf will fetch the bdev for test. At that time, vbdev_gpt_detection may not be ready.  In my opinion, we need to make those detection complete within bdev initialization. And currently our bdev_register is synchronous, which can not satisfy the requirements

Patch-set: 9
Label: Code-Review=-1