vhost: added per-virtqueue task pool
Task ring is now single-producer, single-consumer. This is not only a
performance improvement, but a critical bugfix. Previous (global) ring
was multi-producer, single-consumer and was failing miserably with more
than one vhost controller.

This patch also unifies and generalizes vhost blk & scsi task
management.

The vhost subsystem has been kept to keep the size of this patch small.
It will be removed with next patch.

Change-Id: I6e8395597eb431c25e4c16141b251f26855e0c23
Signed-off-by: Dariusz Stojaczyk <dariuszx.stojaczyk@intel.com>