Update patch set 9
Patch Set 9:

If split is specified for a bdev, and gpt module is enabled and the bdev has an SPDK partition, there will be a conflict.  This would be a user error.  In a real world situation, these types of conflicts should not be possible, since metadata such as the partition table guard against two bdev modules claiming the same bdev.

Yes - I will need to consider the asynchronous registration case.  Maybe the vbdev modules need to tell the generic bdev layer when they are done processing the registration for a given bdev.  Then the bdev layer can delay next subsystem initialization until all registrations are processed.  This would give RPC calls a good hook as well.  I will look into it.

Patch-set: 9
Label: Code-Review=0