bdev: track number of outstanding bdev_registered calls
Add a new spdk_vbdev_register_handled() API which vbdevs
can use to notify the generic bdev layer once it is done
tasting a new bdev.  This is especially required for
asynchronous vbdevs like GPT which must issue I/O to the
base bdev.

As part of this patch, add bdev_registered callbacks
for both split and error, which for now only call
this new functions.  Later patches will move code from
the init callback to the bdev_registered callback for these
modules.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: I49f2d012d1675b878bcd23afff427c740c6502c7