unittests.sh: use valgrind if installed
Automatically uses valgrind with the same options as CI does
provided that it is available and CONFIG_ASAN is not set in
CONFIG.local.

NOTE: the default is to use valgrind if available, to
disable set env var on your dev system $novalgrind to
anything you want or use ASAN and set the option in
CONFIG.local.  Unset it to re-enable valgrind on
every run of unittest.sh

Change-Id: I44f8322680eb2608464eb128d49d20dc1d311f0c
Signed-off-by: Paul Luse <paul.e.luse@intel.com>