Update patch set 2
Patch Set 2: Code-Review-1

Address sanitizer is much newer, lower overhead, and integrated directly into the compiler so I think many people are going to opt to use that over valgrind when running tests locally (at least, I am). Forcing people to set an extra environment variable to run the unit tests, when they've already selected to compile with asan enabled in ./configure, is clunky. Let's come up with a more automatic solution.

One way to do it would be to parse CONFIG.local to check if asan is enabled. If it is enabled, don't use valgrind. You should also probably still check for the presence of the 'valgrind' environment variable, if the user wanted to run it with different flags or turn it off even though asan is not enabled.

Patch-set: 2
Label: Code-Review=-1