Update patch set 2
Patch Set 2:

OK, will look at the ASAN suggestion, it already does that for CI tests so just need to do the same check for local runs.  

Wrt your last comment, the patch is already looking to see if there is an existing valgrind var to enable devs to use their own cmd line and it doesn't overwrite it if so.

Patch-set: 2
Label: Code-Review=0