jsonrpc: remove incorrect handling for send() == 0
recv() returns 0 when the remote end of the connection is shut down, but
there is no such rule for send().  Remove the incorrect treatment of
send() returning 0 as an error.  The remaining code will treat a send()
return value of 0 as a non-error condition and will continue to work
correctly.

Change-Id: Ia753b802d95428104a62d1acb13c5fa437add6b4
Signed-off-by: Daniel Verkamp <daniel.verkamp@intel.com>