Update patch set 3
Patch Set 3:

> (2 comments)
 > 
 > Hi Isaac,
 > 
 > Thanks for your patch.  I noted a couple of small nits.  I have a
 > concern about what happens if there are open bdevs to the NVMe
 > controller when this RPC is issued.  I think that probably we
 > should only allow updating firmware if the NVMe controller is not
 > opened by an upper layer protocol (i.e. iSCSI/NVMe-oF/vhost) nor
 > claimed by a virtual bdev module.
 > 
 > What do you think?
 > 
 > If you agree, then some of this code should be moved into
 > blockdev_nvme.c itself.  Then you can open the bdev read/write
 > (based on some of the recent changes we've pushed to master).  If
 > the open fails, it means it has already been opened by an upper
 > layer protocol or claimed by a virtual bdev module.
 > 
 > -Jim

Hi Jim,

This is a good suggestion. So, when there are multiple physical bdevs per controller, we will need to open all of them. If any open fails, then we disallow the update.

I see that 369486 is removing the differentiation between physical and virtual bdevs. Should we use the examine callback to identify physical bdevs?

-Paul

Patch-set: 3