bdev: defer spdk_bdev_unregister when last descriptor is closed
Commit 5533c3d2 deferred io_channel destruction to allow the
current context to unwind.  This opens a case in the hotplug
path, where the io channel will try to be destroyed after
the io_device has been deleted and the bdev's destruct method
invoked.

So the solution is to defer the latter so that it happens
after the io channel is released.

Reported-by: Dariusz Stojaczyk <dariuszx.stojaczyk@intel.com>
Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: I7378bbf80d9d4c155465022b6fe9f0d412d574c3