examples: blobstore hello world
This is ready for prelim review only. Still some cleanup in
makefiles, maybe some extra stuff floating around, etc. For
now looking for:

- in general is this what we're after?
- are there other functions that we should add now or is
this good to start?
- there are a few // TODO comments in here that are really
questions for Ben or whomever
- is use of the SPDK framework correct and efficient?

haven't run this with valgrind or the script chaecker
or anything but have stepped through it all with gdb
and confirmed its doing what I think its supsoed to be
doing

Change-Id: I1662139e6b1fa33c6109006407a11c8c0936ac0c
Signed-off-by: Paul Luse <paul.e.luse@intel.com>