Update patch set 3
Patch Set 3: Code-Review-1

Darek - can you take a look at my patch series that ends here: https://review.gerrithub.io/#/c/370528/. This is an alternate approach to all this deferring stuff that I think is a lot simpler. If my patch series works, we'll go with that. If it doesn't, we can use this one.

Patch-set: 3