scsi/lun: fix hotremove with no io_channels open
When LUN with no io_channels is being hotremoved,
the hotremove poller was being started on lun->lcore,
This field is only valid when there are io_channels
open. It is 0 by default. For example - when user
tries to delete a bdev that's attached to an unused
vhost controller (no io_channels open) the scsi layer
could start a hotremove poller on lcore 0. If SPDK runs
on a different cpuset - not containing CPU0 - SPDK
could even segfault.

Change-Id: I31a363352875d944f220b0296d5cfe570319023d
Signed-off-by: Dariusz Stojaczyk <dariuszx.stojaczyk@intel.com>